+++ .idea/misc.xml
@@ -36,7 +36,7 @@
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_personal.xml" value="0.2798913043478261" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_terms.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_configuration.xml" value="0.1" />
-        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_cs.xml" value="0.33" />
+        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_cs.xml" value="0.25" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_cs_center.xml" value="0.22" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_cs_detail.xml" value="0.24074074074074073" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_cs_list.xml" value="0.22" />
